norm + red error corr + exec error corr

+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: blaurent <blaurent@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/17 11:47:18 by blaurent          #+#    #+#             */
-/*   Updated: 2023/01/11 17:56:45 by blaurent         ###   ########.fr       */
+/*   Updated: 2023/01/12 14:54:16 by blaurent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -105,7 +105,8 @@ int	main(int ac, char **av, char **envp)
 			{
 				printcmd(c);
 				execute(c, &d);
-				unlink(d.filename);
+				if (d.filename)
+					unlink(d.filename);
 			}
 			free_cmd(c);
 		}
